requirements.txt
pandas==3.0.1
python-dotenv==1.2.2
kafka-python-ng==2.2.3
yfinance==1.2.0